services/auth-service/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.128.0
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.39.0
python-dotenv==1.0.1
pydantic-settings==2.7.0
sqlalchemy==2.0.36
email-validator==2.2.0
python-jose==3.4.0
passlib[bcrypt]==1.7.4
